foundation/storage/src/test/kotlin/com/pingidentity/storage/MemoryStorageTest.kt
/*
 * Copyright (c) 2024. PingIdentity. All rights reserved.
 *
 * This software may be modified and distributed under the terms
 * of the MIT license. See the LICENSE file for details.
 */

package com.pingidentity.storage

import kotlinx.coroutines.test.runTest
import kotlin.test.AfterTest
import kotlin.test.Test
import kotlin.test.assertEquals
import kotlin.test.assertNotNull
import kotlin.test.assertNull

class MemoryStorageTest {
    @AfterTest
    fun tearDown() =
        runTest {
        }

    @Test
    fun testDataStore() =
        runTest {
            val storage = MemoryStorage<Data>()
            storage.save(Data(1, "test"))
            val storedData = storage.get()
            assertEquals(1, storedData!!.a)
            assertEquals("test", storedData.b)
        }

    @Test
    fun testMultipleData() =
        runTest {
            val storage = MemoryStorage<List<Data>>()
            val dataList = listOf(Data(1, "test1"), Data(2, "test2"))
            storage.save(dataList)
            val storedData = storage.get()
            assertEquals(dataList, storedData)
        }

    @Test
    fun testDeleteData() =
        runTest {
            val storage = MemoryStorage<Data>()
            val data = Data(1, "test")
            storage.save(data)
            storage.delete()
            val storedData = storage.get()
            assertEquals(null, storedData)
        }

    @Test
    fun testDifferentDataObjectsWithSameDataStore() =
        runTest {
            val storageData = MemoryStorage<Data>()
            val storageData2 = MemoryStorage<Data2>()

            val data = Data(1, "test")
            val data2 = Data2(2, "test1")

            storageData.save(data)
            storageData2.save(data2)

            val storedData = storageData.get()
            val storedData2 = storageData2.get()

            assertEquals(data, storedData)
            assertEquals(data2, storedData2)

            storageData.delete()
            assertNull(storageData.get())
            assertNotNull(storageData2.get())
        }
}
